+++ finagle-zipkin-core/src/test/scala/com/twitter/finagle/zipkin/core/RawZipkinTracerTest.scala
@@ -1,7 +1,6 @@
 package com.twitter.finagle.zipkin.core
 
 import com.twitter.conversions.DurationOps._
-import com.twitter.finagle.stats.NullStatsReceiver
 import com.twitter.finagle.tracing._
 import com.twitter.util._
 import java.net.{InetAddress, InetSocketAddress}
@@ -11,7 +10,7 @@ class RawZipkinTracerTest extends FunSuite {
 
   val traceId = TraceId(Some(SpanId(123)), Some(SpanId(123)), SpanId(123), None, Flags().setDebug)
 
-  class FakeRawZipkinTracer extends RawZipkinTracer(NullStatsReceiver) {
+  class FakeRawZipkinTracer extends RawZipkinTracer {
     var spans: Seq[Span] = Seq.empty
     override def sendSpans(xs: Seq[Span]): Future[Unit] = {
       spans ++= xs